Feat: 리사이클러뷰 추가,삭제 작동 - #209
