+++ app/src/main/java/com/hsu/mapapp/map/MapViewModel.kt
@@ -6,7 +6,9 @@ import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.google.android.material.snackbar.Snackbar
+import com.google.firebase.auth.FirebaseAuth
 import com.google.firebase.auth.ktx.auth
+import com.google.firebase.firestore.FieldValue
 import com.google.firebase.firestore.FirebaseFirestore
 import com.google.firebase.ktx.Firebase
 import kotlinx.coroutines.launch
@@ -15,32 +17,63 @@ import kotlin.collections.ArrayList
 import kotlin.jvm.internal.MagicApiIntrinsics
 
 class MapViewModel : ViewModel() {
-    val mapRepository = MapRepository()
     var mapData = ArrayList<MapItemList>()
     var mapLiveData: MutableLiveData<ArrayList<MapItemList>> =
         MutableLiveData()
 
+    private var firestore: FirebaseFirestore? = null
+    private var uid = FirebaseAuth.getInstance().currentUser?.uid
+
     init {
-        mapData = mapRepository.getData()
-        mapLiveData.postValue(mapData)
-        println("mapViewModel: $mapData")
+        mapLiveData = fetch()
     }
 
-    fun addMap(mapItem: MapItemList) {
-        mapData.clear()
-        mapData = mapRepository.addData(mapItem)
-        mapLiveData.postValue(mapData)
+    fun fetch(): MutableLiveData<ArrayList<MapItemList>> {
+        firestore = FirebaseFirestore.getInstance()
+        val listData: ArrayList<MapItemList> = arrayListOf()
+        val myRef = firestore?.collection("users")?.document("$uid")
+        myRef!!.get().addOnSuccessListener { document ->
+            if (document.get("mapList") != null) {
+                val mapList: ArrayList<Map<String, String>> =
+                    document.get("mapList") as ArrayList<Map<String, String>>
+                for (map in mapList) {
+                    listData.add(
+                        MapItemList(
+                            map["mapTitle"].toString(),
+                            map["previewImage"].toString(),
+                            map["mapSort"].toString()
+                        )
+                    )
+                }
+            }
+            mapLiveData.postValue(listData)
+        }
+        return mapLiveData
     }
 
-    fun deleteMap(pos: Int) {
-        mapData.clear()
-        mapData = mapRepository.deleteData(pos)
-        mapLiveData.postValue(mapData)
+    fun add(mapItem: MapItemList) {
+        firestore = FirebaseFirestore.getInstance()
+        val myRef = firestore?.collection("users")?.document("$uid")
+        myRef!!.get().addOnSuccessListener { document ->
+            myRef.update("mapList", FieldValue.arrayUnion(mapItem))
+            mapLiveData.postValue(fetch().value)
+        }
     }
 
-    fun editMapTitle(pos: Int, title: String) {
-        mapData.clear()
-        mapData = mapRepository.editMapTitle(pos, title)
-        mapLiveData.postValue(mapData)
+    fun delete(pos: Int) {
+        firestore = FirebaseFirestore.getInstance()
+        val myRef = firestore?.collection("users")?.document("$uid")
+        myRef!!.get().addOnSuccessListener { document ->
+            if (document.get("mapList") != null) {
+                val mapList: ArrayList<Map<String, String>> =
+                    document.get("mapList") as ArrayList<Map<String, String>>
+                myRef.update("mapList", FieldValue.arrayRemove(mapList[pos]))
+                mapList.removeAt(pos)
+                mapLiveData.postValue(fetch().value)
+            }
+        }
+    }
+
+    fun editTitle(pos: Int, title: String) {
     }
 }
